A - Contest Result
There was a contest with N problems. The i-th (1≤i≤N) problem was worth Ai points. Snuke took part in this contest and solved M problems: the B1-th, B2-th, …, and BM-th ones. Find his total score.

Here, the total score is defined as the sum of the points for the problems he solved.