B - Qual B
There were N contestants in the qualification round of a programming contest. All contestants got distinct ranks.You are given a length-N string S, which represents whether the contestants want to participate in the final round or not. Specifically, if the i-th character of S is o, the contestant ranked i-th in the qualification wants to participate in the final; if the i-th character of S is x, the contestant ranked i-th in the qualification does not want to participate in the final.
Among those who want to participate in the final, K contestants with the smallest ranks advance to the final.

Print a string T of length N that satisfies the following conditions:

if the contestant ranked i-th in the qualification advances to the final, the i-th character of T is o;
if the contestant ranked i-th in the qualification does not advance to the final, the i-th character of T is x.